Review how things are.
Too bad Silex was abandoned.  I'm still amazed at how lean it is.
We're reading from files like pigs at each request.
This is not acceptable, but it's how it is.
A quick solution would be to cache the culture as a PHP array
and let the engine handle the RAM optimizations.